Create 2 col video layout
