+++ src/components/homepage/ReleaseVideo.astro
@@ -1,0 +1,15 @@
+---
+import ReleaseVideoPlaceholder from "../../assets/img/audacity-placeholder.webp";
+import PromoProductVideoPlaceholder from "../../assets/img/waves-denoiz-plugin.webp";
+import FeaturedVideo from "../video/FeaturedVideo";
+---
+
+    <section class="bg-gray-900">
+      <div class="grid md:grid-cols-2 max-w-screen-2xl px-6 sm:px-12 mx-auto py-8 gap-8 ">
+        <FeaturedVideo title="Our latest release" label="Master channel, new realtime effects & more!" placeholderImage={ReleaseVideoPlaceholder} videoURL="https://www.youtube-nocookie.com/embed/f5TXPUOFH6A?autoplay=1"/>
+        <FeaturedVideo title="Waves Denoiz Plugin" label="A new generation of noise removal that outpeforms traditional plugins." placeholderImage={PromoProductVideoPlaceholder} videoURL="https://www.youtube-nocookie.com/embed/f5TXPUOFH6A?autoplay=1" CTA ctaText="Get 50% off"/>
+      </div>
+    </section>
+  
+
+
