+++ src/pages/index.astro
@@ -1,7 +1,7 @@
 ---
 import BlogPosts from "../components/homepage/BlogPosts.astro";
 import HeroBanner from "../components/homepage/HeroBanner.astro";
-import ReleaseVideo from "../components/homepage/ReleaseVideo";
+import ReleaseVideo from "../components/homepage/ReleaseVideo.astro";
 import BaseLayout from "../layouts/BaseLayout.astro";
 import ShareYourSounds from "../components/homepage/ShareYourSounds";
 import FeatureCards from "../components/homepage/FeatureCards.astro";
@@ -12,7 +12,7 @@ import RecordAnything from "../components/homepage/RecordAnything.astro";
 
 <BaseLayout title="Audacity ® | Free Audio editor, recorder, music making and more!">
   <HeroBanner />
-  <ReleaseVideo client:load/>
+  <ReleaseVideo />
   <RecordAnything />
   <ShareYourSounds client:load/>
   <FeatureCards />
