+++ tailwind.config.cjs
@@ -4,7 +4,8 @@ module.exports = {
   theme: {
     extend: {},
     screens: {
-      xs: "320px",
+      xxs: "320px",
+      xs: "480px",
       sm: "640px",
       md: "768px",
       lg: "1024px",
